Fixed login bug. Closes #436
